+++ Changelog.md
@@ -23,6 +23,7 @@
 * Give missing sleep resistance to demon lords
 * Lower possibility of destroying statues with spells with higher difficulty
 * Demons, undead and spirits are resistant against dehydration
+* Differentiate the visibility range of monsters
 * Emphasise function key in inventory help line
 * Limit inventory and message dialogue width